requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
-e git+https://github.com/openai/CLIP.git@8a665a683d791ed3491fedadcb3c91878f9eb78d#egg=clip
configparser==5.0.2
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
fsspec==0.8.7
ftfy==5.9
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
patchify==0.2.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1615224089586/work
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.17
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.2.5
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
sentry-sdk==1.0.0
shortuuid==1.0.1
six @ file:///tmp/build/80754af9/six_1605205327372/work
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.7.1
torch-tb-profiler==0.1.0
torchmetrics==0.2.0
torchvision==0.8.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.4
wandb==0.10.23
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
